+++ src/main/java/org/springframework/beans/factory/support/AbstractAutowireCapableBeanFactory.java
@@ -1,7 +1,11 @@
 package org.springframework.beans.factory.support;
 
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.beans.factory.BeansException;
 import org.springframework.beans.factory.config.BeanDefinition;
+
+import java.lang.reflect.Constructor;
 
 /**
  * 实现默认bean创建的抽象bean工厂超类
@@ -11,15 +15,42 @@ import org.springframework.beans.factory.config.BeanDefinition;
  * @project: sbs-small-spring
  */
 public abstract class AbstractAutowireCapableBeanFactory extends AbstractBeanFactory {
+
+  @Getter @Setter
+  private InstantiationStrategy instantiationStrategy = new CglibSubclassingInstantiationStrategy();
+
   @Override
-  protected Object createBean(String beanName, BeanDefinition beanDefinition) {
+  protected Object createBean(String beanName, BeanDefinition beanDefinition, Object[] args) {
     Object bean;
     try {
-      bean = beanDefinition.getBeanClass().getDeclaredConstructor().newInstance();
+      bean = createBeanInstance(beanDefinition, beanName, args);
     } catch (Exception e) {
       throw new BeansException("Instantiation of bean failed", e);
     }
+
     registerSingleton(beanName, bean);
     return bean;
+  }
+
+  /**
+   * 实例化Bean对象
+   *
+   * @param beanDefinition Bean定义
+   * @param beanName bean名称
+   * @param args 参数
+   * @return 结果
+   */
+  protected Object createBeanInstance(
+      BeanDefinition beanDefinition, String beanName, Object[] args) {
+    Constructor constructorToUse = null;
+    Class<?> beanClass = beanDefinition.getBeanClass();
+    Constructor<?>[] declaredConstructors = beanClass.getDeclaredConstructors();
+    for (Constructor ctor : declaredConstructors) {
+      if (null != args && ctor.getParameterTypes().length == args.length) {
+        constructorToUse = ctor;
+        break;
+      }
+    }
+    return getInstantiationStrategy().instantiate(beanDefinition, beanName, constructorToUse, args);
   }
 }