+++ src/main/java/org/springframework/beans/factory/support/SimpleInstantiationStrategy.java
@@ -1,0 +1,39 @@
+package org.springframework.beans.factory.support;
+
+import org.springframework.beans.factory.BeansException;
+import org.springframework.beans.factory.config.BeanDefinition;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.InvocationTargetException;
+
+/**
+ * JDK实例化策略
+ *
+ * @author: DoubleW2w
+ * @date: 2024/9/29
+ * @project: sbs-small-spring
+ */
+public class SimpleInstantiationStrategy implements InstantiationStrategy {
+
+  @Override
+  public Object instantiate(
+      BeanDefinition beanDefinition, String beanName, Constructor ctor, Object[] args)
+      throws BeansException {
+    if (beanDefinition == null) {
+      throw new BeansException("beanDefinition is not exist");
+    }
+    Class clazz = beanDefinition.getBeanClass();
+    try {
+      if (null != ctor) {
+        return clazz.getDeclaredConstructor(ctor.getParameterTypes()).newInstance(args);
+      } else {
+        return clazz.getDeclaredConstructor().newInstance();
+      }
+    } catch (NoSuchMethodException
+        | InstantiationException
+        | IllegalAccessException
+        | InvocationTargetException e) {
+      throw new BeansException("Failed to instantiate [" + clazz.getName() + "]", e);
+    }
+  }
+}