+++ src/test/java/org/springframework/beans/factory/HelloService.java
@@ -6,6 +6,16 @@ package org.springframework.beans.factory;
  * @project: sbs-small-spring
  */
 public class HelloService {
+
+  private String name;
+
+  public HelloService(String name) {
+    this.name = name;
+  }
+
+  public HelloService() {
+  }
+
   public String sayHello() {
     System.out.println("hello");
     return "hello";